+++ app/src/main/java/com/dlab/sinsungo/viewmodel/ShoppingViewModel.kt
@@ -4,11 +4,8 @@ import android.util.Log
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.dlab.sinsungo.IngredientRepository
 import com.dlab.sinsungo.data.model.Shopping
-import com.dlab.sinsungo.data.repository.login.LoginRepository
-import com.dlab.sinsungo.data.repository.shopping.ShoppingRepository
-import kotlinx.coroutines.CoroutineScope
+import com.dlab.sinsungo.data.repository.ShoppingRepository
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
@@ -21,7 +18,7 @@ class ShoppingViewModel : ViewModel() {
     fun requestShopping(newShopping: Shopping) {
         viewModelScope.launch(Dispatchers.IO) {
             try {
-                ShoppingRepository.shopping(newShopping).let {
+                ShoppingRepository.setShopping(newShopping).let {
                     if (it.isSuccessful) {
                         it.body()?.let { res ->
                             withContext(Dispatchers.Main) {
